feat: Rename Retry to SchedulingCoordinator
The object formerly known as Retry actually calls the scheduler
in way to have jobs scheduled afterwards with proper retry settings.
