+++ Tests/Functional/SchedulingCoordinator/ExponentialRetryTest.php
@@ -1,10 +1,10 @@
 <?php
 declare(strict_types=1);
 
-namespace Netlogix\JobQueue\Scheduled\Tests\Functional\Retry;
+namespace Netlogix\JobQueue\Scheduled\Tests\Functional\SchedulingCoordinator;
 
 use Netlogix\JobQueue\Scheduled\Domain\Model\ScheduledJob;
-use Netlogix\JobQueue\Scheduled\Domain\Retry;
+use Netlogix\JobQueue\Scheduled\Domain\SchedulingCoordinator;
 use Netlogix\JobQueue\Scheduled\Domain\Scheduler;
 
 class ExponentialRetryTest extends TestCase
@@ -24,7 +24,7 @@ class ExponentialRetryTest extends TestCase
             'some-claim'
         );
 
-        $retry = new Retry($this->scheduler);
+        $retry = new SchedulingCoordinator($this->scheduler);
         $retry->injectQueueManager($this->queueManager([
             'scheduledJobs' => [
                 'backoffStrategy' => 'exponential',
@@ -53,7 +53,7 @@ class ExponentialRetryTest extends TestCase
             $incarnation
         );
 
-        $retry = new Retry($this->scheduler);
+        $retry = new SchedulingCoordinator($this->scheduler);
         $retry->injectQueueManager($this->queueManager([
             'scheduledJobs' => [
                 'backoffStrategy' => 'exponential',
@@ -82,7 +82,7 @@ class ExponentialRetryTest extends TestCase
             $incarnation
         );
 
-        $retry = new Retry($this->scheduler);
+        $retry = new SchedulingCoordinator($this->scheduler);
         $retry->injectQueueManager($this->queueManager([
             'scheduledJobs' => [
                 'backoffStrategy' => 'exponential',
@@ -115,7 +115,7 @@ class ExponentialRetryTest extends TestCase
             $incarnation
         );
 
-        $retry = new Retry($this->scheduler);
+        $retry = new SchedulingCoordinator($this->scheduler);
         $retry->injectQueueManager($this->queueManager([
             'scheduledJobs' => [
                 'backoffStrategy' => 'exponential',
@@ -173,7 +173,7 @@ class ExponentialRetryTest extends TestCase
             'some-claim'
         );
 
-        $retry = new Retry($this->scheduler);
+        $retry = new SchedulingCoordinator($this->scheduler);
         $retry->injectQueueManager($this->queueManager([
             'scheduledJobs' => [
                 'backoffStrategy' => 'exponential',
